Split interactions types
