+++ packages/types/src/discordeno.ts
@@ -2,8 +2,6 @@ import type {
   AutoModerationActionType,
   AutoModerationEventTypes,
   AutoModerationTriggerTypes,
-  DiscordApplicationCommandOption,
-  DiscordApplicationCommandOptionChoice,
   DiscordApplicationEventWebhookStatus,
   DiscordApplicationIntegrationType,
   DiscordAttachment,
@@ -15,8 +13,6 @@ import type {
   DiscordGuildOnboardingMode,
   DiscordGuildOnboardingPrompt,
   DiscordInstallParams,
-  DiscordInteractionContextType,
-  DiscordInteractionEntryPointCommandHandlerType,
   DiscordMessageReferenceType,
   DiscordPollAnswer,
   DiscordPollLayoutType,
@@ -25,23 +21,30 @@ import type {
   DiscordRole,
   DiscordScheduledEventRecurrenceRule,
 } from './discord.js'
+import type {
+  ApplicationCommandTypes,
+  ButtonStyles,
+  DiscordApplicationCommandOption,
+  DiscordApplicationCommandOptionChoice,
+  DiscordInteractionContextType,
+  DiscordInteractionEntryPointCommandHandlerType,
+  InteractionResponseTypes,
+  MessageComponentTypes,
+  TextStyles,
+} from './discord/interactions.js'
 import type { DiscordWebhookEventType } from './discord/webhookEvents.js'
 import type {
   AllowedMentionsTypes,
-  ApplicationCommandTypes,
   ApplicationFlags,
   AuditLogEvents,
   BigString,
-  ButtonStyles,
   Camelize,
   ChannelTypes,
   DefaultMessageNotificationLevels,
   ExplicitContentFilterLevels,
   ForumLayout,
   GuildFeatures,
-  InteractionResponseTypes,
   Localization,
-  MessageComponentTypes,
   MessageFlags,
   OverwriteTypes,
   PermissionStrings,
@@ -51,7 +54,6 @@ import type {
   SortOrderTypes,
   SystemChannelFlags,
   TargetTypes,
-  TextStyles,
   VerificationLevels,
   VideoQualityModes,
 } from './shared.js'
