+++ packages/types/src/index.ts
@@ -1,5 +1,6 @@
 export * from './discord.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
+export * from './discord/interactions.js'
 export * from './discordeno.js'
 export * from './shared.js'
